# api/swagger/marketplace.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Marketplace API",
    "version": "1.0"
  },
  "host": "marketplace.dev.videocoin.network",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/account": {
      "get": {
        "summary": "Get account info",
        "operationId": "GetAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AccountResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "tags": [
          "Proifle"
        ]
      },
      "put": {
        "summary": "Update account info",
        "operationId": "UpdateAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AccountResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateAccountRequest"
            }
          }
        ],
        "tags": [
          "Proifle"
        ]
      }
    },
...
    "/api/v1/creators": {
      "get": {
        "summary": "Get list of creators",
        "operationId": "GetCreators",
        "security": [],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AccountsResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "integer"
          }
        ],
        "tags": [
          "Creators"
        ]
      }
    },
...
    "/api/v1/creators/{id}": {
      "get": {
        "summary": "Get detail info about creator",
        "operationId": "GetCreator",
        "security": [],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AccountResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number",
            "format": "integer"
          }
        ],
        "tags": [
          "Creators"
        ]
      }
    },
    "/api/v1/my/assets": {
      "get": {
        "summary": "Get list of my assets",
        "operationId": "GetMyAssets",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AssetsResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "tags": [
          "Assets"
        ]
      }
    },
    "/api/v1/spotlight/assets/featured": {
      "get": {
        "summary": "Get list of featured assets",
        "operationId": "GetSpotlightFeaturedAssets",
        "security": [],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AssetsResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "tags": [
          "Spotlight"
        ]
      }
    },
...
    "/api/v1/spotlight/creators/featured": {
      "get": {
        "summary": "Get list of featured creators",
        "operationId": "GetSpotlightFeaturedCreators",
        "security": [],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/AccountsResponse"
            }
          },
          "401": {
            "description": "Returned when the user is not logged in.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "example": {
                "message": "Not found",
                "fields": null
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "integer"
          }
        ],
        "tags": [
          "Spotlight"
        ]
      }
    }
  },
  "definitions": {
...
    "UpdateAccountRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "custom_url": {
          "type": "string"
        },
        "yt_username": {
          "type": "string"
        },
        "image_data": {
          "type": "string"
        },
        "cover_data": {
          "type": "string"
        }
      }
    },
...
    "UserResponse": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "cover_url": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "custom_url": {
          "type": "string"
        },
        "yt_username": {
          "type": "string"
        }
      }
    },
    "AccountResponse": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "profile_img_url": {
          "type": "string"
        },
        "is_verified": {
          "type": "boolean"
        },
        "user": {
          "$ref": "#/definitions/UserResponse"
        }
      }
    },
    "AccountsResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AccountResponse"
          }
        },
        "total_count": {
          "type": "number",
          "format": "integer"
        },
        "count": {
          "type": "number",
          "format": "integer"
        },
        "prev": {
          "type": "boolean",
          "format": "boolean"
        },
        "next": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "AssetContractResponse": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "asset_contract_type": {
          "type": "string"
        },
        "schema_name": {
          "type": "string"
        },
        "symbol": {
          "type": "string"
        },
        "buyer_fee_basis_points": {
          "type": "number"
        },
        "seller_fee_basis_points": {
          "type": "number"
        },
        "opensea_buyer_fee_basis_points": {
          "type": "number"
        },
        "opensea_seller_fee_basis_points": {
          "type": "number"
        },
        "dev_buyer_fee_basis_points": {
          "type": "number"
        },
        "dev_seller_fee_basis_points": {
          "type": "number"
        }
      }
    },
    "AssetResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "format": "integer"
        },
        "token_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "content_type": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "thumbnail_url": {
          "type": "string"
        },
        "preview_url": {
          "type": "string"
        },
        "encrypted_url": {
          "type": "string"
        },
        "qr_url": {
          "type": "string"
        },
        "ipfs_url": {
          "type": "string"
        },
        "ipfs_thumbnail_url": {
          "type": "string"
        },
        "ipfs_encrypted_url": {
          "type": "string"
        },
        "yt_video_id": {
          "type": "string"
        },
        "drm_key": {
          "type": "string"
        },
        "on_sale": {
          "type": "boolean"
        },
        "sold": {
          "type": "boolean"
        },
        "instant_sale_price": {
          "type": "number",
          "format": "float"
        },
        "owner": {
          "$ref": "#/definitions/AccountResponse"
        },
        "asset_contract": {
          "$ref": "#/definitions/AssetContractResponse"
        }
      }
    },
    "AssetsResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AssetResponse"
          }
        },
        "total_count": {
          "type": "number",
          "format": "integer"
        },
        "count": {
          "type": "number",
          "format": "integer"
        },
        "prev": {
          "type": "boolean",
          "format": "boolean"
        },
        "next": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}